# services/product/docs/swagger.yaml
definitions:
  model.ProductDto:
    properties:
      available:
        type: boolean
      name:
        type: string
      price:
        type: number
    required:
    - available
    - name
    - price
    type: object
  model.ProductResponse:
    properties:
      available:
        type: boolean
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    required:
    - available
    - id
    - name
    - price
    type: object
host: localhost:1323
info:
  contact: {}
  description: This is the product service
  title: Product Service
  version: "2.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: Get all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ProductResponse'
            type: array
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Get all products
      tags:
      - product
    post:
      consumes:
      - application/json
      description: Create product
      parameters:
      - description: The input product struct
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/model.ProductDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProductResponse'
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Create product
      tags:
      - product
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Get products by id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProductResponse'
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Get products by id
      tags:
      - product
swagger: "2.0"